fix: change int to bigint

+++ src/main/resources/db/changelog/changeset/0001__init_batch_request_items_table.sql
@@ -3,7 +3,7 @@ CREATE TABLE IF NOT EXISTS batch_request_items (
     name VARCHAR(255) NOT NULL,
     xml_content TEXT,
     status SMALLINT NOT NULL,
-    batch_request_id INT NOT NULL,
+    batch_request_id BIGINT NOT NULL,
 
     CONSTRAINT fk_batch_request_items_batch_requests
     FOREIGN KEY (batch_request_id)
